+++ src/elements/table-of-contents-toc_element/update.js
@@ -1,0 +1,1 @@
+instance.data.renderTableOfContents(properties, context);